rss_parser::parse() writes the feed to cache
This makes cache::internalize_rssfeed() the single source of rss_feed
objects, where previously we could get one either from cache or
rss_parser. There is no difference at the moment, but once we start
using database IDs instead of item's GUIDs, this will become important.

The change shouldn't affect Newsboat at all—the only place where we got
a feed from rss_parser was in controller, and we immediately saved it to
the cache.

Tests had to be updated:

* no mo calling cache::externalize_rssfeed() right after
  rss_parser::parse();

* cache.cpp:692 - since feeds are automatically externalized right after
  parsing, we need to first parse *without* max-items set to check how
  many items there are in the feed;

* cache.cpp:777 - by the time we call set_unread_nowrite(), the item is
  already in the cache, and its unread status won't be updated unless we
  set override_unread as well;

* cache.cpp:810 - same as before. I had to rely on items being unread by
  default, otherwise I would have to use set_override_unread() in the
  setup, which kinda defeats the purpose of the test;

* rss.cpp:293 - items are sorted when they're being internalized, so
  I had to sort the checks too.